+++ app/mailers/application_mailer.rb
@@ -6,13 +6,6 @@ class ApplicationMailer < ActionMailer::Base
 
   before_action :set_title
 
-  # ensure that emails are all in English for the moment
-  def make_bootstrap_mail_en(*args)
-    I18n.with_locale(:en) do
-      make_bootstrap_mail(*args)
-    end
-  end
-
   def set_title
     @title = params[:title] || ""
   end
