+++ app/views/devise/mailer/confirmation_instructions_content.html.erb
@@ -1,0 +1,30 @@
+<h1 class="mb-3">
+  <%= t('devise.mailer.confirmation_instructions.please_confirm_your_energy_sparks_account') %>
+</h1>
+
+<p><%= t('devise.mailer.confirmation_instructions.welcome', resource_name: @resource.name) %>!</p>
+
+<div class="s-3"></div>
+<% if @resource.school.present? %>
+  <p><%= t('devise.mailer.confirmation_instructions.user_account_created_for_you_for', resource_school_name: @resource.school.name) %>.</p>
+<% else %>
+  <p><%= t('devise.mailer.confirmation_instructions.user_account_created_for_you_on') %>.</p>
+<% end %>
+<div class="s-3"></div>
+<p>
+
+<p><%= t('devise.mailer.confirmation_instructions.before_you_begin') %></p>
+
+<%= link_to t('devise.mailer.confirmation_instructions.confirm_my_account'), confirmation_url(@resource, confirmation_token: @token), class: 'btn btn-primary mb-3 mt-3' %>
+
+<div class="s-3"></div>
+
+<p>
+  <%= t('devise.mailer.confirmation_instructions.once_confirmed_message') %>
+</p>
+
+<div class="s-3"></div>
+
+<p>
+  <%= t('devise.mailer.confirmation_instructions.questions_or_concerns_please_contact') %> <%= t('common.application') %> <%= t('devise.mailer.confirmation_instructions.team_by_replying_to_this_email') %>.
+</p>
